LayoutTests/webgpu/whlsl-checker-should-set-type-of-read-modify-write-variables-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS resultsFloat32Array[0] is 42
